moved everything to classes (except main and network)
